requirements.txt
streamlit
pandas
openpyxl
plotly
pydeck
streamlit_folium
folium
Pillow
numpy
matplotlib
Image